# Faster RLP encoding of some types