--- a/kai/storage/kvstore/accessors.go
+++ b/kai/storage/kvstore/accessors.go
@@ -252,7 +252,7 @@ func DeleteCanonicalHash(db kaidb.KeyValueWriter, number uint64) {
 }
 
 // ReadBlockInfo retrieves blockReward, gasUsed and all the transaction receipts belonging to a block.
-func ReadBlockInfo(db kaidb.Reader, hash common.Hash, number uint64) *types.BlockInfo {
+func ReadBlockInfo(db kaidb.Reader, hash common.Hash, number uint64, config *configs.ChainConfig) *types.BlockInfo {
 	// Retrieve the flattened receipt slice
 	data, _ := db.Get(blockInfoKey(number, hash))
 	if len(data) == 0 {
@@ -261,6 +261,15 @@ func ReadBlockInfo(db kaidb.Reader, hash common.Hash, number uint64) *types.Bloc
 	blockInfo := &types.BlockInfo{}
 	if err := rlp.DecodeBytes(data, &blockInfo); err != nil {
 		log.Error("Invalid receipt array RLP", "hash", hash, "err", err)
+		return nil
+	}
+	block := ReadBlock(db, number)
+	if block.Transactions() == nil {
+		log.Error("Missing body but have receipt", "hash", hash, "height", number)
+		return nil
+	}
+	if err := blockInfo.Receipts.DeriveFields(config, hash, number, block.Transactions()); err != nil {
+		log.Error("Failed to derive block receipts fields", "hash", hash, "height", number, "err", err)
 		return nil
 	}
 	return blockInfo
@@ -361,7 +370,7 @@ func ReadReceipt(db kaidb.Reader, hash common.Hash) (*types.Receipt, common.Hash
 	if blockHash == (common.Hash{}) {
 		return nil, common.Hash{}, 0, 0
 	}
-	blockInfo := ReadBlockInfo(db, blockHash, blockHeight)
+	blockInfo := ReadBlockInfo(db, blockHash, blockHeight, nil)
 	if len(blockInfo.Receipts) <= int(receiptIndex) {
 		log.Error("Receipt refereced missing", "number", blockHeight, "hash", blockHash, "index", receiptIndex)
 		return nil, common.Hash{}, 0, 0