--- a/lib/rlp/doc.go
+++ b/lib/rlp/doc.go
@@ -1,3 +1,5 @@
+// From go-ethereum/rlp/doc.go
+
 // Copyright 2014 The go-ethereum Authors
 // This file is part of the go-ethereum library.
 //