--- a/lib/rlp/encbuffer.go
+++ b/lib/rlp/encbuffer.go
@@ -1,3 +1,5 @@
+// From go-ethereum/rlp/encbuffer.go
+
 package rlp
 
 import (