--- a/lib/rlp/rlpgen/gen.go
+++ b/lib/rlp/rlpgen/gen.go
@@ -1,3 +1,5 @@
+// From go-ethereum/rlp/rlpgen/gen.go
+
 package main
 
 import (