--- a/lib/rlp/rlpgen/types.go
+++ b/lib/rlp/rlpgen/types.go
@@ -1,3 +1,5 @@
+// From go-ethereum/rlp/rlpgen/types.go
+
 package main
 
 import (