- [X] Update RLP lib: allow `nil`, `nilString`, `nilList` and `optional` RLP tags in structs; optimize and cleanup changes for `rlp.Stream`.
- [X] Reduced allocs from 6 to 2 and improved speed of `EncodeBig()` function
- [X] Implement `RLPgen` tool to generate `EncodeRLP` (and `DecodeRLP`) method implementations from a struct definition
- [x] Faster RLP encoding of `Acccount`, `ReceiptForStorage` and `BlockInfo`
- [x] Update test units of RLP lib.
- [x] Benchmark receipts and block info RLP encoding
- [x] Ensure backwards compatibility, running stably on devnet at https://github.com/kardiachain/go-kardia/commit/e4353de9ad22fc2fd359fdd440513b191b50b3d9
```
name                              old time/op    new time/op    delta
EncodeRLP/blockInfo-16               971ns ±20%     251ns ± 0%   -74.17%  (p=0.000 n=10+10)
EncodeRLP/receipt-for-storage-16     430ns ±18%     126ns ± 0%   -70.76%  (p=0.000 n=10+10)
EncodeRLP/receipt-full-16            372ns ±11%     357ns ±10%      ~     (p=0.165 n=10+10)
EncodeRLP/transaction-16             257ns ± 1%     268ns ± 0%    +3.94%  (p=0.000 n=10+9)

name                              old speed      new speed      delta
EncodeRLP/blockInfo-16             624MB/s ±24%  2394MB/s ± 0%  +283.78%  (p=0.000 n=10+10)
EncodeRLP/receipt-for-storage-16   764MB/s ±17%  2585MB/s ± 0%  +238.48%  (p=0.000 n=10+10)
EncodeRLP/receipt-full-16          729MB/s ±12%   757MB/s ±11%      ~     (p=0.165 n=10+10)
EncodeRLP/transaction-16           396MB/s ± 1%   381MB/s ± 0%    -3.79%  (p=0.000 n=10+9)

name                              old alloc/op   new alloc/op   delta
EncodeRLP/blockInfo-16                968B ± 0%       40B ± 0%   -95.87%  (p=0.000 n=10+10)
EncodeRLP/receipt-for-storage-16      384B ± 0%        0B       -100.00%  (p=0.000 n=10+10)
EncodeRLP/receipt-full-16             320B ± 0%      320B ± 0%      ~     (all equal)
EncodeRLP/transaction-16             0.00B          0.00B           ~     (all equal)

name                              old allocs/op  new allocs/op  delta
EncodeRLP/blockInfo-16                4.00 ± 0%      2.00 ± 0%   -50.00%  (p=0.000 n=10+10)
EncodeRLP/receipt-for-storage-16      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
EncodeRLP/receipt-full-16             1.00 ± 0%      1.00 ± 0%      ~     (all equal)
EncodeRLP/transaction-16              0.00           0.00           ~     (all equal)
```